Fixed custom field options types and simplified useCustomFieldOptionsBulk
