+++ front/app/api/custom_field_options/__mocks__/useCustomFieldOptions.ts
@@ -8,8 +8,12 @@ export const customFieldOptionsData: ICustomFieldOptionData[] = [
       key: 'key',
       title_multiloc: { en: 'A mock option 1' },
       ordering: 1,
+      other: false,
       created_at: '2020-01-01T00:00:00.000Z',
       updated_at: '2020-01-01T00:00:00.000Z',
+    },
+    relationships: {
+      image: {},
     },
   },
   {
@@ -19,8 +23,12 @@ export const customFieldOptionsData: ICustomFieldOptionData[] = [
       key: 'key2',
       title_multiloc: { en: 'A mock option2' },
       ordering: 2,
+      other: false,
       created_at: '2020-01-01T00:00:00.000Z',
       updated_at: '2020-01-01T00:00:00.000Z',
+    },
+    relationships: {
+      image: {},
     },
   },
 ];
