+++ front/app/api/custom_field_options/types.ts
@@ -1,12 +1,10 @@
-import { Multiloc } from 'typings';
+import { IRelationship, Multiloc } from 'typings';
 
 import { Keys } from 'utils/cl-react-query/types';
 
 import customFieldOptionsKeys from './keys';
 
-export type CustomFieldOptionsKeys = Keys<
-  typeof customFieldOptionsKeys
->;
+export type CustomFieldOptionsKeys = Keys<typeof customFieldOptionsKeys>;
 
 export interface ICustomFieldOptionData {
   id: string;
@@ -15,8 +13,13 @@ export interface ICustomFieldOptionData {
     key: string;
     title_multiloc: Multiloc;
     ordering: number;
+    other: boolean;
+    temp_id?: string;
     created_at: string;
     updated_at: string;
+  };
+  relationships: {
+    image: { data?: IRelationship };
   };
 }
